+++ package.json
@@ -11,7 +11,10 @@
     "build-ts": "tsc",
     "watch-ts": "tsc -w",
     "dev:test": "jest --config=jest.config.json --watchAll",
+    "!test": "------- run self-test with jest ----------------------",
     "test": "jest --config=jest.config.json",
+    "test:dev": "LS=1 jest --config=jest.config.json --watchAll",
+    "test:watch": "LS=1 jest --config=jest.config.json --watchAll",
     "doc": "npm run doc:html && open dist/docs/index.html",
     "doc:html": "typedoc src/libs --exclude **/*.spec.ts --target ES6 --mode file --out dist/docs",
     "doc:publish": "npm run doc:html && gh-pages -m \"docs(gh-pages): publish gh-pages via typedoc\" -d dist/docs",
