+++ src/libs/index.ts
@@ -15,11 +15,13 @@
  * import { CoinValidator } from 'lemon-todaq-lib';
  * ```
  *
- * @author steve@lemoncloud.io
- * @date   2019-06-12
- * @copyright (C) lemoncloud.io 2019 - All Rights Reserved.
+ * @author      Steve Jung <steve@lemoncloud.io>
+ * @date        2019-06-12 initial version
+ * @date        2021-01-22 optimzied and cleanup
+ *
+ * @copyright (C) 2021 LemonCloud Co Ltd. - All Rights Reserved.
  */
-export * from './coin.validator';
-export * from './crypt.helper';
-export * from './date.helper';
-export * from './trans.polling';
+export * from './coin-validator';
+export * from './crypt-helper';
+export * from './date-helper';
+export * from './trans-polling';
