+++ tests/libs/coin-validator.spec.ts
